fix: removing columns
