--- src/pages/User/User.props.tsx
@@ -1,6 +1,0 @@
-import { User } from '../../shared/interfaces';
-
-export const fakeUser: Partial<User> = {
-  name: 'User 1',
-  login: 'aa@bb.cc',
-};
